replace token with plain password auth, finish user fetch

+++ server/server.js
@@ -1,7 +1,6 @@
 const express       = require('express');
 const cors          = require('cors');
 const MongoClient   = require('mongodb').MongoClient;
-const crypto        = require('crypto');
 const cookieParser  = require('cookie-parser');
 const bodyParser    = require('body-parser');
 
@@ -9,27 +8,11 @@ const mongoUrl = 'mongodb://localhost:27017';
 /* TODO put this in .env file with a sane password before deploying
  * oh man, I hope I won't forget this...
  */
-const mainPassword = '1234';
+const adminPass = Buffer.from('1234').toString('base64'); /* updating and deleting */
+const userPass  = Buffer.from('1234').toString('base64'); /* registering */
 
 const app = express();
 var db;
-
-const validateAuth = async (req, res, next) => {
-    next();
-    return;
-
-    const token = req.cookies.authToken;
-    if (token) {
-        const tokens = db.collection('authTokens');
-        const found = await tokens.findOne({ token: token });
-        if (found) {
-            next();
-            return;
-        }
-    }
-
-    res.status(400).send('You are not authenticated.');
-}
 
 app.use(cookieParser());
 app.use(cors());
@@ -40,57 +23,46 @@ app.get('/', (req, res) => {
     res.send('Root sweet root');
 });;
 
-/* authenticate for main page access */
-app.post('/post/authenticate', (req, res) => {
-    const password = req.body.password;
-
-    if (!password) {
-        res.status(400).send('No password specified.');
-        return;
-    }
-
-    if (mainPassword === password) {
-        const token = crypto.randomBytes(16).toString('hex');
-        res.cookie('authToken', token);
-
-        const tokenCollection = db.collection('authTokens');
-        tokenCollection.insertOne({ token: token, date: Date.now() });
-
-        res.send('Authenticated.');
-        return;
-    }
-
-    res.status(400).send('Wrong password.');
-});
-
 /* get users(s) */
-app.get('/get/users', validateAuth, async (req, res) => {
+app.get('/get/users', async (req, res) => {
+    const passInput = req.headers.authorization.split(' ')[1];
+    if (passInput != adminPass) {
+        res.status(400).send('Error: Wrong password.');
+        return;
+    }
+
     const users = db.collection('users');
     const userList = await users.find(req.query).toArray();
 
-    res.json(userList);
+    res.json({ users: userList});
 });
 
 /* delete user */
-app.delete('/delete/user/:email', validateAuth, async (req, res) => {
+app.delete('/delete/user/:email', async (req, res) => {
+    const passInput = req.headers.authorization.split(' ')[1];
+    if (passInput != adminPass) {
+        res.status(400).send('Error: Wrong password.');
+        return;
+    }
+
     const email = req.params.email;
     const users = db.collection('users');
+
     users.deleteOne({ email }, (err, obj) => {
         if (err) {
-            res.status(400).send('Something went wrong.');
+            res.status(400).send('Error: Something went wrong.');
             return;
         }
 
         if (obj.deletedCount > 0)
-            res.send('User deleted.');
+            res.send('Success: User deleted.');
         else
-            res.status(400).send('This user does not exist.');
+            res.status(400).send('Error: This user does not exist.');
     });
 });
 
 /* create and update users */
-app.post('/post/user', validateAuth, async (req, res) => {
-    console.log(req.body);
+app.post('/post/user', async (req, res) => {
     const name      = req.body.name;
     const faculty   = req.body.faculty;
     const email     = req.body.email;
@@ -98,7 +70,7 @@ app.post('/post/user', validateAuth, async (req, res) => {
     const count     = req.body.count;
 
     if (!email) {
-        res.status(400).send('No email specified.');
+        res.status(400).send('Error: No email specified.');
         return;
     }
 
@@ -107,21 +79,44 @@ app.post('/post/user', validateAuth, async (req, res) => {
     const users = db.collection('users');
     const found = await users.findOne({ email });
     if (found) {
+        const passInput = req.headers.authorization.split(' ')[1];
+
         /* user exists => update */
-        users.updateOne({ email }, { $set: user }, (err, res) => {
+        if (passInput != adminPass) {
+            res.status(400).send('Error: Wrong password.');
+            return;
+        }
+
+        users.updateOne({ email }, { $set: user }, (err, _) => {
             if (err) {
                 console.log(err);
-                res.status(400).send('Something went wrong when updating user.');
+                res.status(400).send('Error: Something went wrong when updating user.');
+                return;
             }
+
+            res.send('Success: User updated.');
         });
-        res.send('User updated.');
     } else {
         /* user doesn't exist => create new */
+        const passInput = req.headers.authorization.split(' ')[1];
+
+        if (passInput != adminPass && passInput != userPass) {
+            res.status(400).send('Error: Wrong password.');
+            return;
+        }
+
         if (name && faculty && type && count) {
-            users.insertOne(user);
-            res.send('User inserted.');
+            users.insertOne(user, (err, _) => {
+                if (err) {
+                    console.log(err);
+                    res.status(400).send('Error: Something went wrong when inserting user.');
+                    return;
+                }
+
+                res.send('Success: User inserted.');
+            });
         } else {
-            res.status(400).send('Missing fields when trying to create user.');
+            res.status(400).send('Error: Missing fields when trying to create user.');
         }
     }
 });
